Nest Options class under Plugin class
